+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## [5.24.4](https://github.com/confusedbuffalo/phone-report/compare/v5.24.3...v5.24.4) (2026-07-28)
+
+
+### Bug Fixes
+
+* fix uploaded edits key storing to aid stats display ([7b8f5c8](https://github.com/confusedbuffalo/phone-report/commit/7b8f5c8cbaff269a2645df1b3a426a50a3785362))
+
+## [5.24.3](https://github.com/confusedbuffalo/phone-report/compare/v5.24.2...v5.24.3) (2026-07-25)
+
+
+### Bug Fixes
+
+* update stats on country page when edits have been uploaded ([005c240](https://github.com/confusedbuffalo/phone-report/commit/005c24039d0357d069d73aaa76208cfddd1f0158)), closes [#448](https://github.com/confusedbuffalo/phone-report/issues/448)
+
+## [5.24.2](https://github.com/confusedbuffalo/phone-report/compare/v5.24.1...v5.24.2) (2026-07-22)
+
+
+### Bug Fixes
+
+* fix nearby note message ([e90c8aa](https://github.com/confusedbuffalo/phone-report/commit/e90c8aaf8932991512061d5981a777c0560b34da))
+
 ## [5.24.1](https://github.com/confusedbuffalo/phone-report/compare/v5.24.0...v5.24.1) (2026-07-19)